[refactor] move test file.
