+++ tests/FoundationCreate/Expected/GroupLump/PrefixSampleGroup1Suffix.php
