+++ tests/FoundationCreate/Expected/GroupLump/PrefixSampleGroup2Suffix.php
