+++ tests/FoundationCreate/Expected/Individual/PrefixSample2Suffix.php
