+++ tests/FoundationCreate/Expected/Individual/PrefixSample3Suffix.php
