+++ tests/FoundationCreate/Mock/Resources/individual.blade.php
